Add .aliases incl config alias
These aliases are loaded by .zshrc already, so it was missing from the
repo. Additionally, to use the config repo we need this alias.

Note that the alias config uses git explicitly instead of $(which git).
This helps zsh expand the alias to the command and then apply the
completions for it.