Auto setup remote branch on push
I used to use an alias gpub which would mean something like:

```
git push origin -u head
```

I used this everytime I had created a new branch locally, for
development, and wanted to push this branch to origin.

This meant a lot of switching between git push and this alias.

Git supports a config option push.autoSetupRemote that automatically
assumes that a branch should be created with the same name as refered by
head, if that branch does not already exist on the default remote.