Enable autoStash and autoSquash in git
I've been using autoSquash for a long time in workflows now. I have never
encountered moments where I explicitly did not want to use it. For me
this is fine to enable all the time.

Similarly, I rebase often and sometimes my worktree is dirty. So
autoStash makes a lot of sense for me. In case this leads to conflicts,
I can either resolve the conflicts or find my way back to before the
rebase.