Add git retrycommit alias
This helps when a commit fails and you want to recover the commit
message you wrote. After the commit failed you can simply run `git
retrycommit` to restart writing the commit. It will have the previous
commit message prefilled.

A typical reason that a commit fails is because the gpg key has expired.